Move missingIssueName to shared
