+++ src/applications/appeals/995/tests/validations/issues.unit.spec.js
@@ -5,7 +5,6 @@ import { getDate } from '../../utils/dates';
 import {
   checkIssues,
   maxIssues,
-  missingIssueName,
   maxNameLength,
 } from '../../validations/issues';
 
@@ -179,19 +178,6 @@ describe('selectionRequired', () => {
   it('should show not show an error when an additional issue is selected', () => {
     const errors = { addError: sinon.spy() };
     selectionRequired(errors, _, getData(false, true));
-    expect(errors.addError.called).to.be.false;
-  });
-});
-
-describe('missingIssueName', () => {
-  it('should show an error when a name is missing', () => {
-    const errors = { addError: sinon.spy() };
-    missingIssueName(errors);
-    expect(errors.addError.called).to.be.true;
-  });
-  it('should show an error when a name is missing', () => {
-    const errors = { addError: sinon.spy() };
-    missingIssueName(errors, 'test');
     expect(errors.addError.called).to.be.false;
   });
 });
